requirements.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2023.2
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
